requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-filter==2.2.0
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
idna==2.9
Markdown==3.2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
